fix: zero previous versions when bumping is greater than minor
This fixes a problem in the package when we bump a minor or a major.

For example, when we have to bump a minor.
Bumping a minor from 1.8.3 must be 1.9.0, not 1.9.3.
The same is valid for a major, which in the previous example must be
2.0.0, not 2.9.3.

This also fixes the issue reported on https://github.com/di/bump/issues/15